nautilus-pathbar: Allow label shrinking
Currently the labels of the pathbar set a required size based on its
preferred size. That brings the problem that setting a minimum size of
its preferred size doesn't allow to shrink the label when needed, e.g.
when the windows need to be snapped to a side or when making the window
smaller.
To avoid that, implement a GObject that its preferred size its natural
size, and minimum size of 0, allowing the label to shrink when necesary.

https://bugzilla.gnome.org/show_bug.cgi?id=736792