application: handle open
When some application uses gtk_uri_open, it uses GApplication
open support to call applications which handle opening files.

To be allowed to do that, it's needed to override the open vfunc
of GApplication.

Since we did the change to parse our command line options on ourselves,
it's not longer a problem to use the open vfunc, since it's only
only used when called from dbus open.