view: rename class to NautilusFileView
NautilusView is the abstract class which is subclassed
by the main Nautilus views, such as icon & list views.

It's name, however, is a bit generic now that we want to
have an interface between NautilusWindowSlot and NautilusView.

Change the class name to NautilusFilesView, and update
functions and macros to match that.