view: change show-hidden setting only when value differs
Commit 2e1ac98 introduced the new behavior of persistently
storing the show-hidden setting whenever it changes.

This, however, caused an update cycle where NautilusDirectory
tracked the show-hidden change and invalidates the mime types
of the files, sending the apropriate ::file-changed signals.
When NautilusView receives the signals, it unconditionally
updates the action states, which sets the show-hidden action
again, causing another change on the setting and re-triggering
the NautilusDirectory signals, which are detected by NautilusView
and perpetuating the cycle.

Fix that by only storing the show-hidden setting when it
differs from the current one.