window: remove old hack
NautilusWindow immediately displays the newly opened
tab or window and, for that to happen correctly, an
old hack is present.

This hack, however, added some unecessary complexity to
NautilusWindow code, and is completely obsolete nowadays.
It was needed to fix bug 679640, but seems like the real
cause of the bug was fixed and this hack does nothing,
appearently.

Remove the old hack and lets see if will continue to behave
properly.