view: handle view menu
NautilusToolbar handles the view menu, requiring
direct access to the underlying view inside the
window slot.

Since we're wiping out every access to the underlying
view, we shouldn't access it from NautilusToolbar.

To fix that, makes the view handle the view widget.
Since we're making NautilusWindowSlot a wrapper, add
the necessary properties for it to expose view data
without exposing the view itself.