view: add interface
Nautilus is in the proccess of receiving a places view,
based on GtkFileChooser's one.

To be able to handle that, an abstraction layer is needed
between NautilusFilesView and NautilusWindowSlot, so we factor
out the common data between views.

Add the NautilusView interface, and make NautilusFilesView a
NautilusView implementation. Because of the new way we handle
search on the view side, the search logic is rewritten to match
the new expected behavior.

https://bugzilla.gnome.org/show_bug.cgi?id=753871